requirements.txt
anyio==4.3.0
astor==0.8.1
attrdict==2.0.1
Babel==2.14.0
bce-python-sdk==0.9.6
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
common==0.1.2
contourpy==1.2.1
cssselect==1.2.0
cssutils==2.10.2
cycler==0.10.0
Cython==3.0.10
decorator==5.1.1
distlib==0.3.8
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.13.4
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.51.0
fsspec==2024.3.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hub-sdk==0.0.8
huggingface-hub==0.22.2
idna==2.10
imageio==2.34.0
imgaug==0.4.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.4.1
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==24.0
paddleocr==2.7.3
paddlepaddle==2.2.2
pandas==2.2.2
patch-ng==1.17.4
pdf2docx==0.5.8
pdf2image==1.17.0
pillow==10.3.0
platformdirs==4.2.0
premailer==3.10.0
protobuf==3.20.2
psutil==5.9.8
py-cpuinfo==9.0.0
pybboxes==0.1.6
pyclipper==1.3.0.post5
pycryptodome==3.20.0
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.8.1
rarfile==4.2
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.27
sahi==0.11.15
scikit-image==0.22.0
scipy==1.13.0
seaborn==0.13.2
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sympy==1.12
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tifffile==2024.4.18
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.0.239
ultralyticsplus==0.1.0
urllib3==2.2.1
virtualenv==20.25.3
visualdl==2.5.3
Werkzeug==3.0.2
wget==3.2
zipp==3.18.1
